Additional PR feedback
* Replace "int" return value with "int32_t" in pal_networkchange.cpp.
  socket() returns a file descriptor which should be fine to store in an
  int32_t.
* Add two newlines that were missing at the end of two files.

+++ src/Common/src/Interop/Linux/System.Native/Interop.NetworkChange.cs
@@ -20,12 +20,12 @@ internal static partial class Interop
         public delegate void NetworkChangeEvent(NetworkChangeKind kind);
 
         [DllImport(Libraries.SystemNative)]
-        public static extern IntPtr CreateNetworkChangeListenerSocket();
+        public static extern int CreateNetworkChangeListenerSocket();
 
         [DllImport(Libraries.SystemNative)]
-        public static extern Error CloseNetworkChangeListenerSocket(IntPtr socket);
+        public static extern Error CloseNetworkChangeListenerSocket(int socket);
 
         [DllImport(Libraries.SystemNative)]
-        public static extern NetworkChangeKind ReadSingleEvent(IntPtr socket);
+        public static extern NetworkChangeKind ReadSingleEvent(int socket);
     }
 }
